hash: calculate sizes with size_t instead of longs
... since they return size_t anyway!

+++ lib/cookie.c
@@ -270,11 +270,11 @@ static const char *get_top_domain(const char * const domain, size_t *outlen)
 static size_t cookie_hash_domain(const char *domain, const size_t len)
 {
   const char *end = domain + len;
-  unsigned long h = 5381;
+  size_t h = 5381;
 
   while(domain < end) {
     h += h << 5;
-    h ^= (unsigned long) Curl_raw_toupper(*domain++);
+    h ^= Curl_raw_toupper(*domain++);
   }
 
   return (h % COOKIE_HASH_SIZE);
